# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elapse API",
    "version": "2026-09-06",
    "description": "Per-second billing. Authenticate with your secret key as a bearer token. Every amount is a decimal string in USD."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/v1/checkout/sessions": {
      "post": {
        "operationId": "checkout.sessions.create",
        "tags": [
          "Checkout"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutSession"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The session; send the subscriber to `url`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSession"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No such object in this mode.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Your secret key: `sk_test_…` for test mode, `sk_live_…` for live. Server-side only."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "type",
              "message"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "api_error",
                  "authentication_error",
                  "invalid_request_error",
                  "rate_limit_error",
                  "idempotency_error",
                  "not_found"
                ]
              },
              "message": {
                "type": "string"
              },
              "code": {
                "type": "string",
                "description": "Machine-readable reason, when there is one."
              },
              "param": {
                "type": "string",
                "description": "The request field at fault, when there is one."
              }
            }
          }
        }
      },
...
      "CheckoutSession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cs_3fT8kLm2Qp9RxV"
          },
          "object": {
            "type": "string",
            "enum": [
              "checkout.session"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "complete",
              "expired"
            ]
          },
          "url": {
            "type": "string",
            "description": "Send the subscriber here."
          },
          "livemode": {
            "type": "boolean"
          },
          "created": {
            "type": "integer"
          },
          "expires_at": {
            "type": "integer",
            "description": "Unix seconds; 24 h after creation."
          },
          "success_url": {
            "type": "string"
          },
          "cancel_url": {
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "merchant": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "logo_url": {
                "type": "string",
                "nullable": true
              },
              "accent": {
                "type": "string",
                "nullable": true
              },
              "support_url": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "name",
              "logo_url",
              "accent",
              "support_url"
            ]
          },
          "customer": {
            "type": "string",
            "nullable": true,
            "description": "cus_ id once the subscriber has signed in."
          },
          "subscription": {
            "type": "string",
            "nullable": true,
            "description": "sub_ id once prepared; retrieve it for status."
          },
          "max_duration_seconds": {
            "type": "integer",
            "nullable": true
          },
          "max_escrow_usd": {
            "type": "string",
            "nullable": true,
            "description": "rate × max_duration_seconds, exact decimal."
          }
        },
        "required": [
          "id",
          "object",
          "status",
          "url",
          "livemode",
          "created",
          "expires_at",
          "success_url",
          "cancel_url",
          "product",
          "merchant",
          "customer",
          "subscription",
          "max_duration_seconds",
          "max_escrow_usd"
        ]
      },
      "CreateCheckoutSession": {
        "type": "object",
        "properties": {
          "product": {
            "type": "string",
            "example": "prod_9Xk2mQ1pL0vRsT"
          },
          "success_url": {
            "type": "string",
            "example": "https://acme.test/welcome"
          },
          "cancel_url": {
            "type": "string",
            "example": "https://acme.test/pricing"
          },
          "max_duration_seconds": {
            "type": "integer",
            "minimum": 60,
            "maximum": 2592000,
            "description": "Fix the subscriber's cap. Omit to let them choose on the page (60 s – 30 days)."
          }
        },
        "required": [
          "product",
          "success_url",
          "cancel_url"
        ],
        "additionalProperties": false
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "prod_9Xk2mQ1pL0vRsT"
          },
          "object": {
            "type": "string",
            "enum": [
              "product"
            ]
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "rate_usd_per_second": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d+)?$",
            "example": "0.004",
            "description": "USD per second as a decimal string. Never a float."
          },
          "rate_per_second_wei": {
            "type": "string",
            "example": "4000",
            "description": "Rate in token base units (6 decimals for AUSD)."
          },
          "currency": {
            "type": "string",
            "enum": [
              "ausd"
            ]
          },
          "allow_pause": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          },
          "active_subscriptions": {
            "type": "integer",
            "description": "Running or paused meters on this product."
          },
          "livemode": {
            "type": "boolean"
          },
          "created": {
            "type": "integer",
            "description": "Unix seconds."
          }
        },
        "required": [
          "id",
          "object",
          "name",
          "description",
          "rate_usd_per_second",
          "rate_per_second_wei",
          "currency",
          "allow_pause",
          "active",
          "active_subscriptions",
          "livemode",
          "created"
        ]
      },
...
    }
  }
}